# dashboard/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EvaluationResponse": {
        "description": "An offline measurement on labeled examples, not on live traffic.",
        "properties": {
          "calibration": {
            "items": {
              "$ref": "#/components/schemas/TierCalibration"
            },
            "title": "Calibration",
            "type": "array"
          },
          "cascade": {
            "$ref": "#/components/schemas/CascadeBreakdown"
          },
          "data_source": {
            "anyOf": [
              {
                "enum": [
                  "public",
                  "synthetic",
                  "private"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Source"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "setups": {
            "items": {
              "$ref": "#/components/schemas/EvaluationSetup"
            },
            "title": "Setups",
            "type": "array"
          },
          "split": {
            "title": "Split",
            "type": "string"
          },
          "tiers": {
            "items": {
              "$ref": "#/components/schemas/EvaluationTier"
            },
            "title": "Tiers",
            "type": "array"
          }
        },
        "required": [
          "domain",
          "split",
          "n",
          "data_source",
          "tiers",
          "setups",
          "cascade",
          "calibration"
        ],
        "title": "EvaluationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RequestItem": {
        "properties": {
          "action": {
            "enum": [
              "accepted",
              "human_review"
            ],
            "title": "Action",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "delivery_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Error"
          },
          "delivery_status": {
            "anyOf": [
              {
                "enum": [
                  "pending",
                  "sent",
                  "failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Status"
          },
          "final_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Label"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "review_status": {
            "anyOf": [
              {
                "enum": [
                  "pending",
                  "resolved"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Status"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          }
        },
        "required": [
          "id",
          "created_at",
          "text",
          "label",
          "confidence",
          "tier",
          "action",
          "cost_usd",
          "latency_ms",
          "degraded",
          "review_status",
          "final_label",
          "resolved_at",
          "delivery_status",
          "delivery_error",
          "delivered_at"
        ],
        "title": "RequestItem",
        "type": "object"
      },
...
      "ReviewDecision": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label"
        ],
        "title": "ReviewDecision",
        "type": "object"
      },
      "ReviewItem": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "suggested_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Label"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          }
        },
        "required": [
          "id",
          "created_at",
          "text",
          "suggested_label",
          "confidence",
          "tier"
        ],
        "title": "ReviewItem",
        "type": "object"
      },
      "ReviewResolved": {
        "properties": {
          "final_label": {
            "title": "Final Label",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "status": {
            "const": "resolved",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "final_label"
        ],
        "title": "ReviewResolved",
        "type": "object"
      },
...
      "RouteRequest": {
        "properties": {
          "text": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "RouteRequest",
        "type": "object"
      },
      "RouteResponse": {
        "properties": {
          "action": {
            "enum": [
              "accepted",
              "human_review"
            ],
            "title": "Action",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "request_id": {
            "title": "Request Id",
            "type": "integer"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          }
        },
        "required": [
          "label",
          "confidence",
          "tier",
          "action",
          "cost_usd",
          "latency_ms",
          "degraded",
          "request_id"
        ],
        "title": "RouteResponse",
        "type": "object"
      },
      "StatsResponse": {
        "properties": {
          "accepted_by_tier": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Accepted By Tier",
            "type": "object"
          },
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "degraded": {
            "title": "Degraded",
            "type": "integer"
          },
          "delivery": {
            "$ref": "#/components/schemas/DeliveryCounts"
          },
          "human_review": {
            "title": "Human Review",
            "type": "integer"
          },
          "latency_ms": {
            "$ref": "#/components/schemas/LatencyStats"
          },
          "requests": {
            "title": "Requests",
            "type": "integer"
          },
          "review": {
            "$ref": "#/components/schemas/ReviewCounts"
          },
          "reviewer_agreement": {
            "$ref": "#/components/schemas/ReviewerAgreement"
          }
        },
        "required": [
          "requests",
          "accepted_by_tier",
          "human_review",
          "reviewer_agreement",
          "review",
          "degraded",
          "cost_usd",
          "latency_ms",
          "delivery"
        ],
        "title": "StatsResponse",
        "type": "object"
      },
      "TierCalibration": {
        "properties": {
          "bins": {
            "items": {
              "$ref": "#/components/schemas/CalibrationBin"
            },
            "title": "Bins",
            "type": "array"
          },
          "ece": {
            "description": "Expected calibration error: 0 means the confidence is what it claims to be.",
            "title": "Ece",
            "type": "number"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "tier",
          "ece",
          "bins"
        ],
        "title": "TierCalibration",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RouteIQ",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/evaluation": {
      "get": {
        "description": "Accuracy, cost and calibration measured offline on labeled examples, not on live traffic.",
        "operationId": "get_evaluation",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Get Evaluation"
      }
    },
...
    "/requests": {
      "get": {
        "operationId": "list_requests",
        "parameters": [
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "accepted",
                    "human_review"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "in": "query",
            "name": "tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 50,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          },
          {
            "in": "query",
            "name": "review_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "resolved"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Review Status"
            }
          },
          {
            "in": "query",
            "name": "delivery_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "sent",
                    "failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Delivery Status"
            }
          },
          {
            "in": "query",
            "name": "degraded",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Degraded"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "newest",
              "enum": [
                "newest",
                "oldest"
              ],
              "title": "Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RequestItem"
                  },
                  "title": "Response List Requests",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Total-Count": {
                "description": "Number of requests matching the filters, ignoring limit and offset",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Requests"
      }
    },
    "/review": {
      "get": {
        "operationId": "list_review",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewItem"
                  },
                  "title": "Response List Review",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Total-Count": {
                "description": "Number of requests pending review, ignoring limit",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Review"
      }
    },
    "/review/{request_id}": {
      "post": {
        "operationId": "resolve_review",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResolved"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Review"
      }
    },
    "/route": {
      "post": {
        "operationId": "route_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route Request"
      }
    },
    "/stats": {
      "get": {
        "operationId": "stats",
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stats"
      }
    },
    "/stats/timeseries": {
      "get": {
        "operationId": "timeseries",
        "parameters": [
          {
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "default": "hour",
              "enum": [
                "hour",
                "day"
              ],
              "title": "Bucket",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeseriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Timeseries"
      }
    }
  }
}